fix: v2 전용 Postgres 스키마로 격리 + match_type 컬럼 타입 불일치 수정
실제 Supabase(v1이 지금도 cron으로 쓰고 있는 라이브 프로젝트)에 JDBC로 직접 붙여
Flyway 마이그레이션을 검증하는 과정에서 발견/수정:

- 같은 데이터베이스의 public 스키마에 v1의 matches/match_details 테이블이 이미 존재해
  V2 마이그레이션이 충돌했다 -> v1을 전혀 건드리지 않도록 v2 전용 Postgres 스키마로
  완전히 격리(datasource URL의 currentSchema, spring.flyway.schemas/default-schema,
  hibernate.default_schema 전부 v2로 통일)
- matches/opponent_streaks.match_type을 SMALLINT로 만들었는데 JPA(MatchTypeConverter,
  Integer)는 INTEGER를 기대해 Hibernate 스키마 검증이 실패함 -> INTEGER로 수정

검증: v2 스키마에 V1~V4 마이그레이션 적용 후 ./gradlew :backend:bootRun으로 실제 기동,
GET /actuator/health, /api/v1/users, /api/v1/seasons, /api/v1/visitors/summary 정상 응답 확인
(Postgres -> Flyway -> JPA -> Repository -> Service -> Facade -> Controller 전 계층 e2e 검증)

+++ backend/src/main/resources/db/migration/V2__create_match_and_match_detail.sql
@@ -5,7 +5,7 @@
 CREATE TABLE matches (
     match_id   VARCHAR(64) PRIMARY KEY,
     match_date TIMESTAMPTZ NOT NULL,
-    match_type SMALLINT    NOT NULL,
+    match_type INTEGER     NOT NULL,
     CONSTRAINT ck_matches_match_type CHECK (match_type IN (40, 50))
 );
